+++ core/oevattbeoxviewconfig.php
@@ -50,4 +50,31 @@ class oeVATTBEOxViewConfig extends oeVATTBEOxViewConfig_parent
     {
         return $this->getConfig()->getEdition();
     }
+
+    /**
+     * Check if currently active theme is based on flow.
+     *
+     * @return bool
+     */
+    public function isActiveThemeBasedOnFlow()
+    {
+        return (bool) ('flow' == $this->getActiveTheme()) || ('flow' == $this->getParentThemeId());
+    }
+
+    /**
+     * Getter for parent theme id.
+     *
+     * @return string
+     */
+    protected function getParentThemeId()
+    {
+        $parentThemeId = '';
+        $themeName = $this->getActiveTheme();
+        $theme = oxNew(\OxidEsales\Eshop\Core\Theme::class);
+        $theme->load($themeName);
+        if ($parentTheme = $theme->getParent()) {
+            $parentThemeId = $parentTheme->getId();
+        }
+        return $parentThemeId;
+    }
 }